remove some code duplication
