+++ app/controllers/file_controller.rb
@@ -3,7 +3,7 @@
 ##
 # API for delivering files from stacks
 class FileController < ApplicationController
-  before_action :set_cors_headers, only: [:auth_check]
+  include AuthConcern
 
   rescue_from ActionController::MissingFile do
     render plain: 'File not found', status: :not_found
@@ -29,44 +29,7 @@ class FileController < ApplicationController
     head :ok
   end
 
-  # jsonp response
-  def auth_check
-    # IE 11 and Edge on Windows 10 doesn't request the correct format. So just hardcode
-    # JSON as the return format since that's what we always do.
-    render json: hash_for_auth_check, callback: allowed_params[:callback]
-  end
-
   private
-
-  def set_cors_headers
-    response.headers['Access-Control-Allow-Origin'] = Settings.cors.allow_origin_url
-    response.headers['Access-Control-Allow-Credentials'] = 'true'
-  end
-
-  def allowed_params
-    params.permit(:action, :callback, :id, :file_name, :format)
-  end
-
-  def hash_for_auth_check
-    if can? :access, current_file
-      {
-        status: :success,
-        access_restrictions: {
-          stanford_restricted: current_file.stanford_restricted?,
-          restricted_by_location: current_file.restricted_by_location?,
-          embargoed: current_file.embargoed?,
-          embargo_release_date: current_file.embargo_release_date
-        }
-      }
-    else
-      AuthenticationJson.new(
-        user: current_user,
-        ability: current_ability,
-        file: current_file,
-        auth_url: iiif_auth_api_url
-      )
-    end
-  end
 
   def disposition
     return :attachment if file_params[:download]
@@ -76,6 +39,10 @@ class FileController < ApplicationController
 
   def file_params
     params.permit(:id, :file_name, :download)
+  end
+
+  def allowed?
+    can? :access, current_file
   end
 
   # called when CanCan::AccessDenied error is raised, typically by authorize!
